converted all benchmarks
