+++ samples/sample_benchmark.cpp
@@ -338,112 +338,13 @@ public:
 			g_camera.m_zoom = 25.0f * 0.6f;
 		}
 
-		b2BodyId groundId;
-		{
-			b2BodyDef bodyDef = b2DefaultBodyDef();
-			groundId = b2CreateBody( m_worldId, &bodyDef );
-		}
-
-		{
-			b2BodyDef bodyDef = b2DefaultBodyDef();
-			bodyDef.type = b2_dynamicBody;
-			bodyDef.enableSleep = true;
-			bodyDef.position = { 0.0f, 10.0f };
-			b2BodyId bodyId = b2CreateBody( m_worldId, &bodyDef );
-
-			b2ShapeDef shapeDef = b2DefaultShapeDef();
-			shapeDef.density = 50.0f;
-
-			b2Polygon polygon;
-			polygon = b2MakeOffsetBox( 0.5f, 10.0f, { 10.0f, 0.0f }, b2Rot_identity );
-			b2CreatePolygonShape( bodyId, &shapeDef, &polygon );
-			polygon = b2MakeOffsetBox( 0.5f, 10.0f, { -10.0f, 0.0f }, b2Rot_identity );
-			b2CreatePolygonShape( bodyId, &shapeDef, &polygon );
-			polygon = b2MakeOffsetBox( 10.0f, 0.5f, { 0.0f, 10.0f }, b2Rot_identity );
-			b2CreatePolygonShape( bodyId, &shapeDef, &polygon );
-			polygon = b2MakeOffsetBox( 10.0f, 0.5f, { 0.0f, -10.0f }, b2Rot_identity );
-			b2CreatePolygonShape( bodyId, &shapeDef, &polygon );
-
-			shapeDef.customColor = b2_colorBlueViolet;
-			b2Circle circle = { { 5.0f, 5.0f }, 1.0f };
-			b2CreateCircleShape( bodyId, &shapeDef, &circle );
-			circle = { { 5.0f, -5.0f }, 1.0f };
-			b2CreateCircleShape( bodyId, &shapeDef, &circle );
-			circle = { { -5.0f, -5.0f }, 1.0f };
-			b2CreateCircleShape( bodyId, &shapeDef, &circle );
-			circle = { { -5.0f, 5.0f }, 1.0f };
-			b2CreateCircleShape( bodyId, &shapeDef, &circle );
-
-			// m_motorSpeed = 9.0f;
-			m_motorSpeed = 25.0f;
-
-			b2RevoluteJointDef jd = b2DefaultRevoluteJointDef();
-			jd.bodyIdA = groundId;
-			jd.bodyIdB = bodyId;
-			jd.localAnchorA = { 0.0f, 10.0f };
-			jd.localAnchorB = { 0.0f, 0.0f };
-			jd.referenceAngle = 0.0f;
-			jd.motorSpeed = ( b2_pi / 180.0f ) * m_motorSpeed;
-			jd.maxMotorTorque = 1e8f;
-			jd.enableMotor = true;
-
-			m_jointId = b2CreateRevoluteJoint( m_worldId, &jd );
-		}
-
-		int gridCount = g_sampleDebug ? 20 : 45;
-		b2Polygon polygon = b2MakeBox( 0.125f, 0.125f );
-		b2BodyDef bodyDef = b2DefaultBodyDef();
-		bodyDef.type = b2_dynamicBody;
-		b2ShapeDef shapeDef = b2DefaultShapeDef();
-
-		float y = -0.2f * gridCount + 10.0f;
-		for ( int i = 0; i < gridCount; ++i )
-		{
-			float x = -0.2f * gridCount;
-
-			for ( int j = 0; j < gridCount; ++j )
-			{
-				bodyDef.position = { x, y };
-				b2BodyId bodyId = b2CreateBody( m_worldId, &bodyDef );
-
-				b2CreatePolygonShape( bodyId, &shapeDef, &polygon );
-
-				x += 0.4f;
-			}
-
-			y += 0.4f;
-		}
-	}
-
-	void UpdateUI() override
-	{
-		float height = 60.0f;
-		ImGui::SetNextWindowPos( ImVec2( 10.0f, g_camera.m_height - height - 50.0f ), ImGuiCond_Once );
-		ImGui::SetNextWindowSize( ImVec2( 200.0f, height ) );
-		ImGui::Begin( "Benchmark: Tumbler", nullptr, ImGuiWindowFlags_NoResize );
-		ImGui::PushItemWidth( 120.0f );
-
-		if ( ImGui::SliderFloat( "Speed", &m_motorSpeed, 0.0f, 100.0f, "%.f" ) )
-		{
-			b2RevoluteJoint_SetMotorSpeed( m_jointId, ( b2_pi / 180.0f ) * m_motorSpeed );
-
-			if ( m_motorSpeed > 0.0f )
-			{
-				b2Joint_WakeBodies( m_jointId );
-			}
-		}
-
-		ImGui::PopItemWidth();
-		ImGui::End();
+		CreateTumbler( m_worldId );
 	}
 
 	static Sample* Create( Settings& settings )
 	{
 		return new BenchmarkTumbler( settings );
 	}
-
-	b2JointId m_jointId;
-	float m_motorSpeed;
 };
 
 static int benchmarkTumbler = RegisterSample( "Benchmark", "Tumbler", BenchmarkTumbler::Create );
@@ -648,49 +549,10 @@ public:
 		{
 			g_camera.m_center = { 0.0f, 50.0f };
 			g_camera.m_zoom = 25.0f * 2.2f;
+			settings.enableSleep = false;
 		}
 
-#ifdef NDEBUG
-		int baseCount = 100;
-#else
-		int baseCount = 40;
-#endif
-
-		{
-			b2BodyDef bodyDef = b2DefaultBodyDef();
-			bodyDef.position = { 0.0f, -1.0f };
-			b2BodyId groundId = b2CreateBody( m_worldId, &bodyDef );
-
-			b2Polygon box = b2MakeBox( 100.0f, 1.0f );
-			b2ShapeDef shapeDef = b2DefaultShapeDef();
-			b2CreatePolygonShape( groundId, &shapeDef, &box );
-		}
-
-		b2BodyDef bodyDef = b2DefaultBodyDef();
-		bodyDef.type = b2_dynamicBody;
-
-		b2ShapeDef shapeDef = b2DefaultShapeDef();
-		shapeDef.density = 1.0f;
-
-		float h = 0.5f;
-		b2Polygon box = b2MakeRoundedBox( h - 0.05f, h - 0.05f, 0.05f );
-
-		float shift = 1.0f * h;
-
-		for ( int i = 0; i < baseCount; ++i )
-		{
-			float y = ( 2.0f * i + 1.0f ) * shift;
-
-			for ( int j = i; j < baseCount; ++j )
-			{
-				float x = ( i + 1.0f ) * shift + 2.0f * ( j - i ) * shift - h * baseCount;
-
-				bodyDef.position = { x, y };
-
-				b2BodyId bodyId = b2CreateBody( m_worldId, &bodyDef );
-				b2CreatePolygonShape( bodyId, &shapeDef, &box );
-			}
-		}
+		CreateLargePyramid(m_worldId);
 	}
 
 	static Sample* Create( Settings& settings )
@@ -711,149 +573,16 @@ public:
 		{
 			g_camera.m_center = { 16.0f, 110.0f };
 			g_camera.m_zoom = 25.0f * 5.0f;
+			settings.enableSleep = false;
 		}
 
-		m_extent = 0.5f;
-		m_round = 0.0f;
-		m_baseCount = 10;
-		m_rowCount = g_sampleDebug ? 4 : 20;
-		m_columnCount = g_sampleDebug ? 4 : 20;
-		m_groundId = b2_nullBodyId;
-		m_bodyIds = nullptr;
-		m_bodyCount = 0;
-		m_bodyIndex = 0;
-
-		CreateScene();
-	}
-
-	~BenchmarkManyPyramids() override
-	{
-		free( m_bodyIds );
-	}
-
-	void CreatePyramid( float centerX, float baseY )
-	{
-		b2BodyDef bodyDef = b2DefaultBodyDef();
-		bodyDef.type = b2_dynamicBody;
-
-		b2ShapeDef shapeDef = b2DefaultShapeDef();
-		shapeDef.density = 1.0f;
-
-		float h = m_extent - m_round;
-		b2Polygon box = b2MakeRoundedBox( h, h, m_round );
-
-		float shift = 1.0f * h;
-
-		for ( int i = 0; i < m_baseCount; ++i )
-		{
-			float y = ( 2.0f * i + 1.0f ) * shift + baseY;
-
-			for ( int j = i; j < m_baseCount; ++j )
-			{
-				float x = ( i + 1.0f ) * shift + 2.0f * ( j - i ) * shift + centerX - 0.5f;
-
-				bodyDef.position = { x, y };
-
-				assert( m_bodyIndex < m_bodyCount );
-				m_bodyIds[m_bodyIndex] = b2CreateBody( m_worldId, &bodyDef );
-				b2CreatePolygonShape( m_bodyIds[m_bodyIndex], &shapeDef, &box );
-
-				m_bodyIndex += 1;
-			}
-		}
-	}
-
-	void CreateScene()
-	{
-		if ( B2_IS_NON_NULL( m_groundId ) )
-		{
-			b2DestroyBody( m_groundId );
-		}
-
-		for ( int i = 0; i < m_bodyCount; ++i )
-		{
-			b2DestroyBody( m_bodyIds[i] );
-		}
-
-		free( m_bodyIds );
-
-		m_bodyCount = m_rowCount * m_columnCount * m_baseCount * ( m_baseCount + 1 ) / 2;
-		m_bodyIds = (b2BodyId*)malloc( m_bodyCount * sizeof( b2BodyId ) );
-		m_bodyIndex = 0;
-
-		b2BodyDef bodyDef = b2DefaultBodyDef();
-		m_groundId = b2CreateBody( m_worldId, &bodyDef );
-
-		float groundDeltaY = 2.0f * m_extent * ( m_baseCount + 1.0f );
-		float groundWidth = 2.0f * m_extent * m_columnCount * ( m_baseCount + 1.0f );
-		b2ShapeDef shapeDef = b2DefaultShapeDef();
-
-		float groundY = 0.0f;
-
-		for ( int i = 0; i < m_rowCount; ++i )
-		{
-			// b2Segment segment = {{-0.5f * groundWidth, groundY}, {0.5f * groundWidth, groundY}};
-			b2Segment segment = { { -0.5f * 2.0f * groundWidth, groundY }, { 0.5f * 2.0f * groundWidth, groundY } };
-			b2CreateSegmentShape( m_groundId, &shapeDef, &segment );
-			groundY += groundDeltaY;
-		}
-
-		float baseWidth = 2.0f * m_extent * m_baseCount;
-		float baseY = 0.0f;
-
-		for ( int i = 0; i < m_rowCount; ++i )
-		{
-			for ( int j = 0; j < m_columnCount; ++j )
-			{
-				float centerX = -0.5f * groundWidth + j * ( baseWidth + 2.0f * m_extent ) + m_extent;
-				CreatePyramid( centerX, baseY );
-			}
-
-			baseY += groundDeltaY;
-		}
-
-		assert( m_bodyIndex == m_bodyCount );
-	}
-
-	void UpdateUI() override
-	{
-		float height = 160.0f;
-		ImGui::SetNextWindowPos( ImVec2( 10.0f, g_camera.m_height - height - 50.0f ), ImGuiCond_Once );
-		ImGui::SetNextWindowSize( ImVec2( 200.0f, height ) );
-		ImGui::Begin( "Benchmark: Many Pyramids", nullptr, ImGuiWindowFlags_NoResize );
-		ImGui::PushItemWidth( 100.0f );
-
-		bool changed = false;
-		changed = changed || ImGui::SliderInt( "Row Count", &m_rowCount, 1, 32 );
-		changed = changed || ImGui::SliderInt( "Column Count", &m_columnCount, 1, 32 );
-		changed = changed || ImGui::SliderInt( "Base Count", &m_baseCount, 1, 30 );
-
-		changed = changed || ImGui::SliderFloat( "Round", &m_round, 0.0f, 0.4f, "%.1f" );
-		changed = changed || ImGui::Button( "Reset Scene" );
-
-		if ( changed )
-		{
-			CreateScene();
-		}
-
-		ImGui::PopItemWidth();
-		ImGui::End();
+		CreateManyPyramids( m_worldId );
 	}
 
 	static Sample* Create( Settings& settings )
 	{
 		return new BenchmarkManyPyramids( settings );
 	}
-
-	b2BodyId m_groundId;
-	b2BodyId* m_bodyIds;
-	int m_bodyCount;
-	int m_bodyIndex;
-	int m_baseCount;
-	int m_rowCount;
-	int m_columnCount;
-	float m_round;
-	float m_extent;
 };
 
 static int benchmarkManyPyramids = RegisterSample( "Benchmark", "Many Pyramids", BenchmarkManyPyramids::Create );
@@ -1134,94 +863,16 @@ public:
 		{
 			g_camera.m_center = { 60.0f, -57.0f };
 			g_camera.m_zoom = 25.0f * 2.5f;
+			settings.enableSleep = false;
 		}
 
-		constexpr int N = g_sampleDebug ? 10 : 100;
-
-		// Allocate to avoid huge stack usage
-		b2BodyId* bodies = static_cast<b2BodyId*>( malloc( N * N * sizeof( b2BodyId ) ) );
-		int index = 0;
-
-		b2ShapeDef shapeDef = b2DefaultShapeDef();
-		shapeDef.density = 1.0f;
-		shapeDef.filter.categoryBits = 2;
-		shapeDef.filter.maskBits = ~2u;
-
-		b2Circle circle = { { 0.0f, 0.0f }, 0.4f };
-
-		b2RevoluteJointDef jd = b2DefaultRevoluteJointDef();
-		b2BodyDef bodyDef = b2DefaultBodyDef();
-
-		for ( int k = 0; k < N; ++k )
-		{
-			for ( int i = 0; i < N; ++i )
-			{
-				float fk = (float)k;
-				float fi = (float)i;
-
-				if ( k >= N / 2 - 3 && k <= N / 2 + 3 && i == 0 )
-				{
-					bodyDef.type = b2_staticBody;
-				}
-				else
-				{
-					bodyDef.type = b2_dynamicBody;
-				}
-
-				bodyDef.position = { fk, -fi };
-
-				b2BodyId body = b2CreateBody( m_worldId, &bodyDef );
-
-				b2CreateCircleShape( body, &shapeDef, &circle );
-
-				if ( i > 0 )
-				{
-					jd.bodyIdA = bodies[index - 1];
-					jd.bodyIdB = body;
-					jd.localAnchorA = { 0.0f, -0.5f };
-					jd.localAnchorB = { 0.0f, 0.5f };
-					b2CreateRevoluteJoint( m_worldId, &jd );
-				}
-
-				if ( k > 0 )
-				{
-					jd.bodyIdA = bodies[index - N];
-					jd.bodyIdB = body;
-					jd.localAnchorA = { 0.5f, 0.0f };
-					jd.localAnchorB = { -0.5f, 0.0f };
-					b2CreateRevoluteJoint( m_worldId, &jd );
-				}
-
-				bodies[index++] = body;
-			}
-		}
-
-		free( bodies );
-
-		m_gravity = 10.0f;
-	}
-
-	void UpdateUI() override
-	{
-		float height = 60.0f;
-		ImGui::SetNextWindowPos( ImVec2( 10.0f, g_camera.m_height - height - 50.0f ), ImGuiCond_Once );
-		ImGui::SetNextWindowSize( ImVec2( 240.0f, height ) );
-		ImGui::Begin( "Benchmark: Joint Grid", nullptr, ImGuiWindowFlags_NoResize );
-
-		if ( ImGui::SliderFloat( "gravity", &m_gravity, 0.0f, 20.0f, "%.1f" ) )
-		{
-			b2World_SetGravity( m_worldId, { 0.0f, -m_gravity } );
-		}
-
-		ImGui::End();
+		CreateJointGrid( m_worldId );
 	}
 
 	static Sample* Create( Settings& settings )
 	{
 		return new BenchmarkJointGrid( settings );
 	}
-
-	float m_gravity;
 };
 
 static int benchmarkJointGridIndex = RegisterSample( "Benchmark", "Joint Grid", BenchmarkJointGrid::Create );
@@ -1238,46 +889,7 @@ public:
 			g_camera.m_zoom = 25.0f * 1.6f;
 		}
 
-		b2World_SetGravity( m_worldId, b2Vec2_zero );
-
-		{
-			b2Polygon box = b2MakeBox( 4.0f, 4.0f );
-
-			b2BodyDef bodyDef = b2DefaultBodyDef();
-			bodyDef.type = b2_dynamicBody;
-			bodyDef.position = { -20.0f, 0.0f };
-			bodyDef.linearVelocity = { 40.0f, 0.0f };
-			b2BodyId bodyId = b2CreateBody( m_worldId, &bodyDef );
-
-			b2ShapeDef shapeDef = b2DefaultShapeDef();
-			shapeDef.density = 8.0f;
-			b2CreatePolygonShape( bodyId, &shapeDef, &box );
-		}
-
-		{
-			float d = 0.4f;
-			b2Polygon box = b2MakeSquare( 0.5f * d );
-
-			b2BodyDef bodyDef = b2DefaultBodyDef();
-			bodyDef.type = b2_dynamicBody;
-			bodyDef.isAwake = false;
-
-			b2ShapeDef shapeDef = b2DefaultShapeDef();
-
-			int columns = g_sampleDebug ? 20 : 120;
-			int rows = g_sampleDebug ? 10 : 80;
-
-			for ( int i = 0; i < columns; ++i )
-			{
-				for ( int j = 0; j < rows; ++j )
-				{
-					bodyDef.position.x = i * d + 30.0f;
-					bodyDef.position.y = ( j - rows / 2.0f ) * d;
-					b2BodyId bodyId = b2CreateBody( m_worldId, &bodyDef );
-					b2CreatePolygonShape( bodyId, &shapeDef, &box );
-				}
-			}
-		}
+		CreateSmash( m_worldId );
 	}
 
 	static Sample* Create( Settings& settings )
@@ -1852,12 +1464,6 @@ static int sampleCast = RegisterSample( "Benchmark", "Cast", BenchmarkCast::Crea
 class BenchmarkSpinner : public Sample
 {
 public:
-	enum
-	{
-		e_count = 3038,
-		e_pointCount = 360,
-	};
-
 	explicit BenchmarkSpinner( Settings& settings )
 		: Sample( settings )
 	{
@@ -1867,196 +1473,13 @@ public:
 			g_camera.m_zoom = 42.0f;
 		}
 
-		b2BodyId groundId;
-		{
-			b2BodyDef bodyDef = b2DefaultBodyDef();
-			groundId = b2CreateBody( m_worldId, &bodyDef );
-
-			b2Vec2 points[e_pointCount];
-
-			b2Rot q = b2MakeRot( -2.0f * b2_pi / e_pointCount );
-			b2Vec2 p = { 40.0f, 0.0f };
-			for ( int i = 0; i < e_pointCount; ++i )
-			{
-				points[i] = { p.x, p.y + 32.0f };
-				p = b2RotateVector( q, p );
-			}
-
-			b2ChainDef chainDef = b2DefaultChainDef();
-			chainDef.points = points;
-			chainDef.count = e_pointCount;
-			chainDef.isLoop = true;
-			chainDef.friction = 0.1f;
-			chainDef.customColor = b2_colorWhite;
-
-			b2CreateChain( groundId, &chainDef );
-		}
-
-		{
-			m_bodyType = b2_dynamicBody;
-
-			b2BodyDef bodyDef = b2DefaultBodyDef();
-			bodyDef.type = m_bodyType;
-			bodyDef.position = { 0.0, 12.0f };
-			bodyDef.enableSleep = false;
-
-			m_spinnerId = b2CreateBody( m_worldId, &bodyDef );
-
-			b2Polygon box = b2MakeRoundedBox( 0.4f, 20.0f, 0.2f );
-			b2ShapeDef shapeDef = b2DefaultShapeDef();
-			shapeDef.friction = 0.0f;
-			shapeDef.customColor = b2_colorAliceBlue;
-			b2CreatePolygonShape( m_spinnerId, &shapeDef, &box );
-
-			m_motorSpeed = 5.0f;
-			m_maxMotorTorque = 40000.0f;
-			b2RevoluteJointDef jointDef = b2DefaultRevoluteJointDef();
-			jointDef.bodyIdA = groundId;
-			jointDef.bodyIdB = m_spinnerId;
-			jointDef.localAnchorA = bodyDef.position;
-			jointDef.enableMotor = true;
-			jointDef.motorSpeed = m_motorSpeed;
-			jointDef.maxMotorTorque = m_maxMotorTorque;
-
-			m_jointId = b2CreateRevoluteJoint( m_worldId, &jointDef );
-		}
-
-		b2Capsule capsule = { { -0.25f, 0.0f }, { 0.25f, 0.0f }, 0.25f };
-		b2Circle circle = { { 0.0f, 0.0f }, 0.35f };
-		b2Polygon square = b2MakeSquare( 0.35f );
-
-		b2BodyDef bodyDef = b2DefaultBodyDef();
-		bodyDef.type = b2_dynamicBody;
-		b2ShapeDef shapeDef = b2DefaultShapeDef();
-		shapeDef.friction = 0.1f;
-		shapeDef.restitution = 0.1f;
-		shapeDef.density = 0.25f;
-
-		int bodyCount = g_sampleDebug ? 499 : e_count;
-
-		float x = -24.0f, y = 2.0f;
-		for ( int i = 0; i < bodyCount; ++i )
-		{
-			bodyDef.position = { x, y };
-			b2BodyId bodyId = b2CreateBody( m_worldId, &bodyDef );
-
-			int remainder = i % 3;
-			if ( remainder == 0 )
-			{
-				// shapeDef.customColor = b2_colorYellow;
-				b2CreateCapsuleShape( bodyId, &shapeDef, &capsule );
-			}
-			else if ( remainder == 1 )
-			{
-				// shapeDef.customColor = b2_colorYellowGreen;
-				b2CreateCircleShape( bodyId, &shapeDef, &circle );
-			}
-			else if ( remainder == 2 )
-			{
-				// shapeDef.customColor = b2_colorGreenYellow;
-				b2CreatePolygonShape( bodyId, &shapeDef, &square );
-			}
-
-			x += 1.0f;
-
-			if ( x > 24.0f )
-			{
-				x = -24.0f;
-				y += 1.0f;
-			}
-		}
-
-		m_acceleration = 0.0f;
-	}
-
-	void UpdateUI() override
-	{
-		float height = 160.0f;
-		ImGui::SetNextWindowPos( ImVec2( 10.0f, g_camera.m_height - height - 50.0f ), ImGuiCond_Once );
-		ImGui::SetNextWindowSize( ImVec2( 240.0f, height ) );
-
-		ImGui::Begin( "Spinner", nullptr, ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoResize );
-
-		if ( ImGui::SliderFloat( "Speed", &m_motorSpeed, -5.0f, 5.0f ) )
-		{
-			b2RevoluteJoint_SetMotorSpeed( m_jointId, m_motorSpeed );
-
-			if ( m_bodyType == b2_kinematicBody )
-			{
-				b2Body_SetAngularVelocity( m_spinnerId, m_motorSpeed );
-			}
-		}
-
-		if ( ImGui::SliderFloat( "Torque", &m_maxMotorTorque, 0.0f, 100000.0f, "%.0f" ) )
-		{
-			b2RevoluteJoint_SetMaxMotorTorque( m_jointId, m_maxMotorTorque );
-		}
-
-		if ( ImGui::RadioButton( "Static", m_bodyType == b2_staticBody ) )
-		{
-			m_bodyType = b2_staticBody;
-			b2Body_SetType( m_spinnerId, b2_staticBody );
-		}
-
-		if ( ImGui::RadioButton( "Kinematic", m_bodyType == b2_kinematicBody ) )
-		{
-			m_bodyType = b2_kinematicBody;
-			b2Body_SetType( m_spinnerId, b2_kinematicBody );
-			b2Body_SetAngularVelocity( m_spinnerId, m_motorSpeed );
-		}
-
-		if ( ImGui::RadioButton( "Dynamic", m_bodyType == b2_dynamicBody ) )
-		{
-			m_bodyType = b2_dynamicBody;
-			b2Body_SetType( m_spinnerId, b2_dynamicBody );
-		}
-
-		ImGui::End();
-	}
-
-	void Step( Settings& settings ) override
-	{
-		if ( glfwGetKey( g_mainWindow, GLFW_KEY_A ) == GLFW_PRESS )
-		{
-			m_acceleration = -0.2f;
-		}
-
-		if ( glfwGetKey( g_mainWindow, GLFW_KEY_S ) == GLFW_PRESS )
-		{
-			m_acceleration = 0.0f;
-			m_motorSpeed = 0.0f;
-		}
-
-		if ( glfwGetKey( g_mainWindow, GLFW_KEY_D ) == GLFW_PRESS )
-		{
-			m_acceleration = 0.2f;
-		}
-
-		if ( settings.hertz > 0.0f )
-		{
-			m_motorSpeed = b2ClampFloat( m_motorSpeed + m_acceleration / settings.hertz, -5.0f, 5.0f );
-			b2RevoluteJoint_SetMotorSpeed( m_jointId, m_motorSpeed );
-
-			if ( m_bodyType == b2_kinematicBody )
-			{
-				b2Body_SetAngularVelocity( m_spinnerId, m_motorSpeed );
-			}
-		}
-
-		Sample::Step( settings );
+		CreateSpinner( m_worldId );
 	}
 
 	static Sample* Create( Settings& settings )
 	{
 		return new BenchmarkSpinner( settings );
 	}
-
-	b2BodyId m_spinnerId;
-	b2JointId m_jointId;
-	float m_acceleration;
-	float m_motorSpeed;
-	float m_maxMotorTorque;
-	b2BodyType m_bodyType;
 };
 
 static int sampleSpinner = RegisterSample( "Benchmark", "Spinner", BenchmarkSpinner::Create );
@@ -2071,6 +1494,7 @@ public:
 		{
 			g_camera.m_center = { 0.0f, 110.0f };
 			g_camera.m_zoom = 125.0f;
+			settings.enableSleep = true;
 		}
 
 		settings.drawJoints = false;
